removed chaining to load categories
Previously we were fetching loading a new category only after the categories at the top of current category is completed. This result in a dependency to receive a callback from LoaderHelper in order to proceed to the next category. Instead now we are using a simple delay to load categories one after one and all categories are fetched together.